evaluation/requirements.txt
protobuf==5.29.6
python-dotenv==1.2.2
requests==2.33.1
google-generativeai==0.8.6
